Define m_secmarker_skipped as an integer type
There's no reason to treat this this as a double, since it
represents a human-readable data value that is only meaningful
as an integer. In doing so we write cleaner audit logs and save
a small amount of space.

+++ headers/modsecurity/rules.h
@@ -80,7 +80,7 @@ class Rules : public RulesProperties {
 
  private:
     int m_referenceCount;
-    double m_secmarker_skipped;
+    uint8_t m_secmarker_skipped;
 };
 
 #endif
